// src/types/resources/SubscriptionTopic.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCanonical } from '../simpleTypes/Canonical';
import { TDateTime } from '../simpleTypes/DateTime';
import { TContactDetail } from '../partials/ContactDetail';
import { TMarkdown } from '../simpleTypes/Markdown';
import { TUsageContext } from '../partials/UsageContext';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TDate } from '../simpleTypes/Date';
import { TPeriod } from '../partials/Period';
import { TSubscriptionTopicResourceTrigger } from '../partials/SubscriptionTopicResourceTrigger';
import { TSubscriptionTopicEventTrigger } from '../partials/SubscriptionTopicEventTrigger';
import { TSubscriptionTopicCanFilterBy } from '../partials/SubscriptionTopicCanFilterBy';
import { TSubscriptionTopicNotificationShape } from '../partials/SubscriptionTopicNotificationShape';

export type TSubscriptionTopic = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  url: TUri;
  identifier?: TIdentifier[];
  version?: string;
  title?: string;
  derivedFrom?: TCanonical[];
  status: string;
  experimental?: boolean;
  date?: TDateTime;
  publisher?: string;
  contact?: TContactDetail[];
  description?: TMarkdown;
  useContext?: TUsageContext[];
  jurisdiction?: TCodeableConcept[];
  purpose?: TMarkdown;
  copyright?: TMarkdown;
  approvalDate?: TDate;
  lastReviewDate?: TDate;
  effectivePeriod?: TPeriod;
  resourceTrigger?: TSubscriptionTopicResourceTrigger[];
  eventTrigger?: TSubscriptionTopicEventTrigger[];
  canFilterBy?: TSubscriptionTopicCanFilterBy[];
  notificationShape?: TSubscriptionTopicNotificationShape[];
};